refactor(mcp/write): #463 extract CLUSTER_A detection helpers into write_detection.py (STRATEGY_S1)
First of four cluster extractions in STRATEGY_S1 (parent #459) decomposing
the 4887-LOC write.py god-object. CLUSTER_A is the pure-text detection /
W-code emitter cluster — no AST, no I/O, no WriteTool state.

Moved verbatim into new module src/octave_mcp/mcp/write_detection.py:

  Functions (13):
    _build_literal_zone_line_set
    _build_holographic_line_set
    _all_section_marks_quoted
    _auto_quote_section_refs_in_values        (138 LOC lenient repair)
    _detect_unquoted_section_in_values        (audit walk)
    _build_annotation_protected_ranges
    _is_in_protected_range
    _count_brackets_outside_protected
    _detect_annotation_too_long               (W_ANNOTATION_TOO_LONG)
    _w_snake_case_blob_token_excluded
    _w_snake_case_blob_token_triggers
    _w_snake_case_blob_build_warning
    _detect_snake_case_blob                   (177 LOC, W_SNAKE_CASE_BLOB)

  Plus their exclusively-owned constants moved alongside:
    W_UNQUOTED_SECTION_IN_VALUE,
    W_ANNOTATION_TOO_LONG, _ANNOTATION_TOO_LONG_CHAR_LIMIT,
    _ANNOTATION_TOO_LONG_TOKEN_LIMIT, W_SNAKE_CASE_BLOB,
    _W_SNAKE_CASE_BLOB_* (REASONING_FIELDS, STOPWORDS, ALLCAPS_RE,
    TOKEN_RE, OPENER_RE, BLOCK_OPEN_RE, BULK_CHAR_LIMIT,
    BULK_UNDERSCORE_LIMIT, SEMANTIC_STOPWORD_LIMIT),
    _ANNOTATION_QUALIFIER_RE, _UNQUOTED_SECTION_RE,
    _LITERAL_ZONE_FENCE_RE, _SECTION_REF_TOKEN_RE,
    _SECTION_CONTAINING_TOKEN_RE.

No behaviour change. No signature change. Function bodies byte-identical.

Importer updates (no re-export shim per #459 acceptance):
  - src/octave_mcp/mcp/validate.py        : _detect_snake_case_blob
  - src/octave_mcp/mcp/write.py           : internal-use import of the
    four helpers WriteTool itself still calls (_auto_quote_section_refs_in_values,
    _detect_unquoted_section_in_values, _detect_annotation_too_long,
    _detect_snake_case_blob). This is a real call site, not a forwarding shim.
  - tests/unit/test_gh452_snake_case_blob.py
  - tests/unit/test_gh403_annotation_too_long.py
  - tests/unit/test_pr361_review_fixes.py
  - tests/unit/test_write_tool.py            (26 inline imports updated)
  - tests/integration/test_schema_write_idempotency.py (5 inline imports)

LOC delta:
  src/octave_mcp/mcp/write.py:           4887 -> 3961 (-926)
  src/octave_mcp/mcp/write_detection.py:    0 ->  946 (new)
  net:                                   4887 -> 4907 (+20 — docstring + import header)

Quality gates (all green):
  pytest -q --no-cov : 3614 passed, 11 skipped, 3 xfailed   (delta=0 vs origin/main)
  ruff check         : All checks passed!
  black --check      : 213 files would be left unchanged
  mypy src           : Success: no issues found in 54 source files

Refs: #459 (STRATEGY_S1 parent), #463 (CLUSTER_A this PR)
Next: #464 CLUSTER_B, #465 CLUSTER_D, #466 CLUSTER_C
Precondition for: v1.14.0 hybrid fix (#460)
