+++ tests/unit/test_pr361_review_fixes.py
@@ -16,7 +16,8 @@ import pytest
 
 from octave_mcp.core.grammar.cst import Comment
 from octave_mcp.core.parser import parse_with_warnings
-from octave_mcp.mcp.write import WriteTool, _all_section_marks_quoted, _auto_quote_section_refs_in_values
+from octave_mcp.mcp.write import WriteTool
+from octave_mcp.mcp.write_detection import _all_section_marks_quoted, _auto_quote_section_refs_in_values
 
 
 class TestIssue1EscapeUnawareAutoQuoteScanner:
